+++ changelog.md
@@ -4,6 +4,8 @@
 * Support for do_verp
 * Fix bug in CRAM-MD5 AUTH
 * Propagate debug output option to SMTP class (@Reblutus)
+* Determine MIME type of attachments automatically
+* Add cross-platform, multibyte-safe pathinfo repalcement and use it
 
 ## Version 5.2.6 (April 11th 2013)
 * Reflect move to PHPMailer GitHub organisation at https://github.com/PHPMailer/PHPMailer